+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 8.06.2018
 
+* [Rust] Code Royale - Count units fn (CodinGame)
 * [Rust] Code Royale - Prioritised building mines (CodinGame)
 
 #### 7.06.2018